+++ server/lib/prosody/config/content.ts
@@ -157,7 +157,11 @@ class ProsodyConfigContent {
       'uptime', // Report how long server has been running
       'ping', // Replies to XMPP pings with pongs
       'bosh', // Enable BOSH clients, aka "Jabber over HTTP"
-      'posix' // POSIX functionality, sends server to background, enables syslog, etc.
+      'posix', // POSIX functionality, sends server to background, enables syslog, etc.
+      'pep', // Enables users to publish their avatar, mood, activity, playing music and more
+      'vcard_legacy' // Conversion between legacy vCard and PEP Avatar, vcard
+      // 'vcard4' // User profiles (stored in PEP)
+
     ])
     this.global.set('modules_disabled', [
       // 'offline' // Store offline messages